Add OrderDetailService and OrderDetailController

+++ Infrastructure/Repositories/IOrderDetailRepository.cs
@@ -6,7 +6,7 @@ namespace Infrastructure.Repositories
     public interface IOrderDetailRepository
     {
         Task<int> CreateAsync(OrderDetail orderDetail);
-        Task<int> UpdateAsync(int id, int orderId, int menuItemId, int quantity, int price, OrderDetailStatus status);
+        Task<int> UpdateAsync(int id, int orderId, int menuItemId, int quantity, decimal price, OrderDetailStatus status);
         Task<int> DeleteAsync(int id);
         Task<OrderDetail?> GetAsync(int id);
         Task<List<OrderDetail>> GetAllAsync();
